# docs/openapi/event-service.openapi.yml
openapi: 3.0.3
info:
  title: BoardHub Event Service API
  version: 0.20.0
  description: API REST dell'event-service BoardHub per controllo dei tavoli, sessioni, ingresso dei giocatori, personaggi, eventi e movimento su griglia.
servers:
  - url: http://localhost:8082
    description: Ambiente locale di sviluppo
...
paths:
...
  /api/v1/admin/tables:
    get:
      tags:
        - Venue
      summary: Elenca l'inventario dei tavoli del locale
      description: Operazione privata e, per impostazione predefinita, accessibile soltanto dal computer del backend.
      security:
        - VenueKey: []
      responses:
        "200":
          description: Stato corrente di tutti i tavoli configurati
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicTableStatusResponse"
        "401":
          description: Credenziale del locale assente, errata o richiesta non locale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/tables/{tablePublicId}/enable:
    post:
      tags:
        - Venue
      summary: Abilita temporaneamente l'avvio di una sessione
      security:
        - VenueKey: []
      parameters:
        - $ref: "#/components/parameters/TablePublicId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableTableRequest"
      responses:
        "200":
          description: Tavolo reclamabile fino alla scadenza indicata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTableStatusResponse"
        "401":
          description: Credenziale del locale non valida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Il tavolo ospita gia una sessione
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/admin/tables/{tablePublicId}/disable:
    post:
      tags:
        - Venue
      summary: Revoca una finestra di avvio non ancora utilizzata
      security:
        - VenueKey: []
      parameters:
        - $ref: "#/components/parameters/TablePublicId"
      responses:
        "200":
          description: Tavolo disabilitato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTableStatusResponse"
  /api/v1/admin/tables/{tablePublicId}/close-session:
    post:
      tags:
        - Venue
      summary: Conclude dal locale la sessione attiva del tavolo
      security:
        - VenueKey: []
      parameters:
        - $ref: "#/components/parameters/TablePublicId"
      responses:
        "200":
          description: Sessione conclusa e tavolo disabilitato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTableStatusResponse"
  /api/v1/public/tables/{tablePublicId}:
    get:
      tags:
        - Tables
      summary: Restituisce lo stato pubblico del tavolo
      description: Distingue un tavolo disabilitato, temporaneamente reclamabile o con sessione attiva. Sono validi i QR dei tavoli configurati da 1 a 8.
      parameters:
        - name: tablePublicId
          in: path
          required: true
          schema:
            type: string
          example: qr-table-04
      responses:
        "200":
          description: Stato pubblico del tavolo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicTableStatusResponse"
        "400":
          description: QR malformato o numero fuori dall'intervallo configurato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/public/tables/{tablePublicId}/active-session:
    get:
      tags:
        - Tables
      summary: Risolve il QR del tavolo nella sessione attiva
      parameters:
        - name: tablePublicId
          in: path
          required: true
          schema:
            type: string
          example: qr-table-04
      responses:
        "200":
          description: Informazioni pubbliche della sessione disponibile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicActiveSessionResponse"
        "404":
          description: Tavolo senza una sessione pubblica attiva
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/public/sessions/{sessionId}/join-requests:
    post:
      tags:
        - Join
      summary: Richiede al DM di entrare nella sessione
      description: La stessa chiave di idempotenza e lo stesso body restituiscono la richiesta gia creata.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          example: session-20260705-001
        - name: Idempotency-Key
          in: header
          required: true
          description: UUID nuovo generato dal client per questa operazione.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJoinRequestRequest"
      responses:
        "201":
          description: Richiesta inserita nella coda del DM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinRequestResponse"
        "409":
          description: Sessione chiusa, richiesta duplicata o chiave riusata con altri dati
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Troppe richieste dallo stesso giocatore
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/public/sessions/{sessionId}/join-requests/{requestId}:
    get:
      tags:
        - Join
      summary: Recupera lo stato della propria richiesta di ingresso
      description: >
        Il dispositivo conserva la chiave casuale usata per creare la richiesta e la presenta
        come credenziale di possesso. Se la richiesta e stata accettata, la risposta include
        il partecipante e la sua credenziale firmata di sessione.
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/RequestId"
        - name: X-BoardHub-Join-Claim
          in: header
          required: true
          description: UUID segreto usato come Idempotency-Key durante la creazione.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Stato corrente della richiesta posseduta dal dispositivo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerJoinStatusResponse"
        "404":
          description: Richiesta inesistente oppure chiave di possesso errata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    VenueKey:
      type: apiKey
      in: header
      name: X-BoardHub-Venue-Key
      description: Credenziale amministrativa del locale, non esposta ai QR o ai client pubblici.
    DmSessionBearer:
      type: http
      scheme: bearer
      bearerFormat: bhd1
      description: Credenziale del Dungeon Master, associata a una sola sessione attiva.
    PlayerBearer:
      type: http
      scheme: bearer
      bearerFormat: bhp1
      description: Credenziale HMAC associata a un partecipante e a una sola sessione attiva.
  parameters:
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      example: session-20260705-001
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TablePublicId:
      name: tablePublicId
      in: path
      required: true
      schema:
        type: string
        pattern: "^qr-table-[0-9]{2}$"
      example: qr-table-04
  schemas:
...
    PublicActiveSessionResponse:
      type: object
      properties:
        tablePublicId:
          type: string
          example: qr-table-04
        tableDisplayName:
          type: string
          example: Tavolo 4
        sessionId:
          type: string
          example: session-20260705-001
        title:
          type: string
          example: Cripta del Re Caduto
        gameType:
          type: string
          example: DND
        publicSummary:
          type: string
          example: Avventura per personaggi di livello 3.
      required: [tablePublicId, tableDisplayName, sessionId, title, gameType, publicSummary]
    PublicTableStatusResponse:
      type: object
      properties:
        tablePublicId:
          type: string
          example: qr-table-04
        tableNumber:
          type: integer
          minimum: 1
          maximum: 8
          example: 4
        tableDisplayName:
          type: string
          example: Tavolo 4
        status:
          type: string
          enum: [DISABLED, CLAIMABLE, IN_SESSION]
        claimExpiresAt:
          type: string
          format: date-time
          nullable: true
          description: Scadenza della finestra di claim, presente soltanto nello stato CLAIMABLE.
        activeSession:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/PublicActiveSessionResponse"
      required: [tablePublicId, tableNumber, tableDisplayName, status]
    EnableTableRequest:
      type: object
      properties:
        durationMinutes:
          type: integer
          minimum: 1
          maximum: 60
          description: Durata opzionale della finestra di avvio.
    CreateJoinRequestRequest:
      type: object
      properties:
        playerReference:
          type: string
          description: Identificativo locale stabile dell'installazione del giocatore.
          example: player-device-01
        displayName:
          type: string
          maxLength: 80
          example: Andrea
      required: [playerReference, displayName]
    JoinRequestResponse:
      type: object
      properties:
        requestId:
          type: string
          format: uuid
        sessionId:
          type: string
        playerReference:
          type: string
        displayName:
          type: string
        status:
          type: string
          enum: [PENDING, ACCEPTED, REJECTED, EXPIRED]
        requestedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
      required: [requestId, sessionId, playerReference, displayName, status, requestedAt, expiresAt]
    ParticipantResponse:
      type: object
      properties:
        participantId:
          type: string
          format: uuid
        sessionId:
          type: string
        role:
          type: string
          enum: [DM, PLAYER]
        displayName:
          type: string
        status:
          type: string
          enum: [ACTIVE, LEFT]
        joinedAt:
          type: string
          format: date-time
      required: [participantId, sessionId, role, displayName, status, joinedAt]
...
    PlayerJoinStatusResponse:
      type: object
      properties:
        request:
          $ref: "#/components/schemas/JoinRequestResponse"
        participant:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/ParticipantResponse"
        accessToken:
          type: string
          nullable: true
          description: Presente soltanto quando la richiesta e ACCEPTED.
      required: [request]
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: SESSION_NOT_FOUND
        message:
          type: string
          example: "Sessione non trovata: session-non-esistente"
      required:
        - code
        - message
...